AEROGEAR-8246 - move voyagerResolver to its own file
